requirements.txt
streamlit==1.39.0
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
openai==1.47.0
tiktoken==0.7.0
python-docx==1.1.2
numpy==1.26.4
scikit-learn==1.5.2
requests==2.32.3
beautifulsoup4==4.12.3
gspread==6.1.2
google-auth>=2.0.0
trafilatura==1.7.0
feedparser==6.0.11   # (optional, for later RSS)
tqdm==4.66.5
beautifulsoup4>=4.12
